Fix LSA

+++ src/control/lsa.zig
@@ -9,7 +9,7 @@ const proto = lib.proto;
 /// principal contra ciclos y el TTL es el cinturon de seguridad del acuerdo.
 pub const DEFAULT_TTL: i32 = 8;
 
-pub fn build(me: []const u8, seq: u32, links: proto.LinkMap) proto.Lsa {
+pub fn build(me: []const u8, seq: i64, links: proto.LinkMap) proto.Lsa {
     return .{
         .origin = me,
         .seq = seq,
